backend/requirements.txt
fastapi>=0.115.0,<1
httpx>=0.27.0,<1
pydantic>=2.8.0,<3
pyyaml>=6.0.0,<7
uvicorn[standard]>=0.30.0,<1
